refactor: move document.dart to core/document
